moving code around, update README
